Refactored arabic calculation to remove lookup table and to add range exceptions.
I actually prefer the lookup table version because its intent is clearer, but its the solution which most people performing the roman numerals kata come up with, and few people try to actually discover the algorithm involved.  This solution still seems to contain repetition, as 9 and 4 are very similar, and 6-8 and 1-3 are also similar, and could possibly be further optimised.